--- a/Cargo.toml
+++ b/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "detached-shell"
-version = "0.1.1"
+version = "0.1.2"
 readme = "README.md"
 edition = "2021"
 authors = ["Kerem Noras <kerem@noras.tech>"]