… Cargo.lock
… PERFORMANCE_SECURITY_IMPROVEMENTS.md
… README.md
… src/pty/spawn.rs
… tests/integration_test.rs
… tests/security_test.rs